moving cogs into components folder
