+++ .gitignore
@@ -4,3 +4,4 @@ bot_token.txt
 /savedata
 /__pycache__
 /downloads/image
+/components/__pycache__
